Split windowed reductions and their gradients into a separate file inside the lax implementation.
