+++ jax/experimental/jet.py
@@ -28,9 +28,7 @@ from jax.tree_util import (register_pytree_node, tree_structure,
 import jax.linear_util as lu
 from jax.interpreters import xla
 from jax.custom_derivatives import custom_jvp_call_jaxpr_p
-from jax._src.lax import lax
-from jax._src.lax import control_flow as lax_control_flow
-from jax._src.lax import fft as lax_fft
+from jax import lax
 
 def jet(fun, primals, series):
   try:
@@ -241,7 +239,7 @@ deflinear(lax.transpose_p)
 deflinear(lax.slice_p)
 deflinear(lax.reduce_sum_p)
 deflinear(lax.reduce_window_sum_p)
-deflinear(lax_fft.fft_p)
+deflinear(lax.fft_p)
 deflinear(xla.device_put_p)
 
 def _cumulative_jet_rule(primals_in, series_in, *, axis: int, reverse: bool,
@@ -249,16 +247,16 @@ def _cumulative_jet_rule(primals_in, series_in, *, axis: int, reverse: bool,
   # Irrespective of backend, we always use the parallel prefix scan
   # implementation when differentiating because reduce_window is not
   # arbitrarily differentiable.
-  return jet(partial(lax_control_flow.associative_scan, combine_fn, axis=axis,
+  return jet(partial(lax.associative_scan, combine_fn, axis=axis,
                      reverse=reverse),
              primals_in, series_in)
 
-deflinear(lax_control_flow.cumsum_p)
-jet_rules[lax_control_flow.cumprod_p] = partial(_cumulative_jet_rule,
+deflinear(lax.cumsum_p)
+jet_rules[lax.cumprod_p] = partial(_cumulative_jet_rule,
                                                 combine_fn=lax.mul)
-jet_rules[lax_control_flow.cummax_p] = partial(_cumulative_jet_rule,
+jet_rules[lax.cummax_p] = partial(_cumulative_jet_rule,
                                                combine_fn=lax.max)
-jet_rules[lax_control_flow.cummin_p] = partial(_cumulative_jet_rule,
+jet_rules[lax.cummin_p] = partial(_cumulative_jet_rule,
                                                combine_fn=lax.min)
