+++ jax/lax/__init__.py
@@ -215,12 +215,6 @@ from jax._src.lax.lax import (
   reduce_precision_p as reduce_precision_p,
   reduce_prod_p as reduce_prod_p,
   reduce_sum_p as reduce_sum_p,
-  reduce_window as reduce_window,
-  reduce_window_max_p as reduce_window_max_p,
-  reduce_window_min_p as reduce_window_min_p,
-  reduce_window_p as reduce_window_p,
-  reduce_window_shape_tuple as reduce_window_shape_tuple,
-  reduce_window_sum_p as reduce_window_sum_p,
   regularized_incomplete_beta_p as regularized_incomplete_beta_p,
   rem as rem,
   rem_p as rem_p,
@@ -247,9 +241,6 @@ from jax._src.lax.lax import (
   scatter_mul_p as scatter_mul_p,
   scatter_p as scatter_p,
   select as select,
-  select_and_gather_add_p as select_and_gather_add_p,
-  select_and_scatter_add_p as select_and_scatter_add_p,
-  select_and_scatter_p as select_and_scatter_p,
   select_p as select_p,
   shift_left as shift_left,
   shift_left_p as shift_left_p,
@@ -295,15 +286,30 @@ from jax._src.lax.lax import (
   xor_p as xor_p,
   zeros_like_array as zeros_like_array,
 )
-from jax._src.lax.lax import (_reduce_sum, _reduce_max, _reduce_min, _reduce_or,
-                  _reduce_and, _reduce_window_sum, _reduce_window_max,
-                  _reduce_window_min, _reduce_window_prod,
-                  _select_and_gather_add,
-                  _select_and_scatter_add, _float, _complex, _input_dtype,
-                  _const, _eq_meet, _broadcasting_select,
-                  _check_user_dtype_supported, _one, _zero,
-                  _upcast_fp16_for_computation, _broadcasting_shape_rule,
-                  _eye, _tri, _delta, _ones, _zeros, _dilate_shape)
+from jax._src.lax.lax import (
+  _reduce_sum, _reduce_max, _reduce_min, _reduce_or, _reduce_and,
+  _float, _complex, _input_dtype,
+  _const, _eq_meet, _broadcasting_select,
+  _check_user_dtype_supported, _one, _zero,
+  _upcast_fp16_for_computation, _broadcasting_shape_rule,
+  _eye, _tri, _delta, _ones, _zeros, _dilate_shape)
+from jax._src.lax.windowed_reductions import (
+  _reduce_window_sum,
+  _reduce_window_max,
+  _reduce_window_min,
+  _reduce_window_prod,
+  _select_and_gather_add,
+  _select_and_scatter_add,
+  reduce_window as reduce_window,
+  reduce_window_max_p as reduce_window_max_p,
+  reduce_window_min_p as reduce_window_min_p,
+  reduce_window_p as reduce_window_p,
+  reduce_window_shape_tuple as reduce_window_shape_tuple,
+  reduce_window_sum_p as reduce_window_sum_p,
+  select_and_gather_add_p as select_and_gather_add_p,
+  select_and_scatter_p as select_and_scatter_p,
+  select_and_scatter_add_p as select_and_scatter_add_p,
+)
 from jax._src.lax.control_flow import (
   associative_scan as associative_scan,
   cond as cond,
